+++ app/Http/Controllers/Auth/ConfirmPasswordController.php
@@ -32,7 +32,7 @@ class ConfirmPasswordController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('auth');
+        $this->middleware('auth.portal');
     }
 
     /**